+++ src/main/java/com/foolsix/fancyenchantments/events/CapabilityEvents.java
@@ -13,6 +13,7 @@ import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Items;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.event.entity.living.LivingEquipmentChangeEvent;
@@ -61,7 +62,9 @@ public class CapabilityEvents {
         if (e.player.tickCount % 20 != 0) {
             e.player.getCapability(ElementStatsCapabilityProvider.PLAYER_ELEMENT_STATS).ifPresent(elementStats -> {
                 if (elementStats.getPoint(AER) >= CONFIG.aerLevelToGetSpeed) {
-                    e.player.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED, 20));
+                    e.player.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED,
+                            20,
+                            elementStats.getPoint(AER) / CONFIG.aerLevelToGetSpeed));
                 }
                 if (elementStats.getPoint(IGNIS) >= CONFIG.ignisLevelToGetFireResistance) {
                     e.player.addEffect(new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 20));
@@ -89,7 +92,6 @@ public class CapabilityEvents {
     @SubscribeEvent
     public void elementPlayerEquipmentChange(LivingEquipmentChangeEvent e) {
         if (!(e.getEntity() instanceof Player player)) return;
-
         player.getCapability(ElementStatsCapabilityProvider.PLAYER_ELEMENT_STATS).ifPresent(elementStats -> {
             for (var entry : e.getFrom().getAllEnchantments().entrySet()) {
                 Element element = Element.getElement(entry.getKey());
@@ -98,6 +100,7 @@ public class CapabilityEvents {
                 }
             }
 
+            if (e.getTo().is(Items.AIR)) return;
             for (var entry : e.getTo().getAllEnchantments().entrySet()) {
                 Element element = Element.getElement(entry.getKey());
                 if (element != null) {